feat: Added ability to dissmiss keyboard
Users made the suggestion to be able to dismiss the keyboard when trying to add a new "Blind Box". The issues was that when trying to enter the name of the new "Blind Box" they keyboard covered the bottom end of the list after the date picker. Due to this, users are not able to adjust the size of the "Blind Box". I added a function to return the keyboard when the "return" key is pressed. This will allow users to dismiss the keyboard and now be able to fully see the list of option when creating a new "Blind Box".